TTOL-9 feat: Navigation 컴포넌트 생성
좌측 네비게이션 부분을 Navigation 컴포넌트로 분리
